refactor(server): relocate static/public route handlers
Move the six static/public HTTP handlers (handleHealth, handlePublicFile,
handlePublicFolder, handlePublicFolderDownload, handleCapabilities,
handleCdpCheck — ~266 lines) out of server.go into static_routes.go.
Pure relocation: same package, no behavior change.

server.go now 8,899 lines (down from 11,738 across all relocations).
fmt/vet clean, cmd/server tests pass.
