+++ agent_go/cmd/server/server.go
@@ -3,16 +3,13 @@ package server
 import (
 	"bytes"
 	"context"
-	"encoding/base64"
 	"encoding/json"
 	"fmt"
 	"hash/fnv"
-	"io"
 	"log"
 	"net"
 	"net/http"
 	_ "net/http/pprof" // Register pprof handlers
-	"net/url"
 	"os"
 	"os/signal"
 	"path/filepath"
@@ -2242,273 +2239,6 @@ func requestLogPath(r *http.Request) string {
 		return r.URL.Path
 	}
 	return r.URL.Path + "?" + r.URL.RawQuery
-}
-
-// Health check endpoint
-func (api *StreamingAPI) handleHealth(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(http.StatusOK)
-
-	// Get current tracing provider
-	tracingProvider := os.Getenv("TRACING_PROVIDER")
-	if tracingProvider == "" {
-		tracingProvider = "noop"
-	}
-
-	json.NewEncoder(w).Encode(map[string]interface{}{
-		"status":  "healthy",
-		"time":    time.Now(),
-		"version": llmtypes.VERSION,
-		"config": map[string]interface{}{"provider": api.config.Provider,
-			"model":            api.config.ModelID,
-			"temperature":      api.config.Temperature,
-			"max_turns":        api.config.MaxTurns,
-			"tracing_provider": tracingProvider,
-		},
-	})
-}
-
-// handlePublicFile serves workspace files via a shareable URL.
-// GET /api/public/file?path=<base64-encoded-filepath>
-func (api *StreamingAPI) handlePublicFile(w http.ResponseWriter, r *http.Request) {
-	encoded := r.URL.Query().Get("path")
-	if encoded == "" {
-		http.Error(w, "path query parameter required", http.StatusBadRequest)
-		return
-	}
-
-	// Decode base64 filepath
-	decoded, err := base64.StdEncoding.DecodeString(encoded)
-	if err != nil {
-		decoded, err = base64.RawURLEncoding.DecodeString(encoded)
-		if err != nil {
-			log.Printf("[PUBLIC-FILE] Failed to decode base64 path: %s, error: %v", encoded, err)
-			http.Error(w, "invalid file path encoding", http.StatusBadRequest)
-			return
-		}
-	}
-	filePath := string(decoded)
-
-	// Use uid from query param (owner's ID for cross-user sharing), fall back to auth context
-	uid := r.URL.Query().Get("uid")
-	if uid == "" {
-		uid = GetUserIDFromContext(r.Context())
-	}
-	log.Printf("[PUBLIC-FILE] Serving file: %s for user: %s", filePath, uid)
-
-	// URL-encode each path segment for the workspace API
-	segments := strings.Split(filePath, "/")
-	for i, seg := range segments {
-		segments[i] = url.PathEscape(seg)
-	}
-	encodedPath := strings.Join(segments, "/")
-
-	wsURL := getWorkspaceAPIURL() + "/api/documents/" + encodedPath + "/raw"
-	log.Printf("[PUBLIC-FILE] Proxying to workspace: %s", wsURL)
-
-	req, err := http.NewRequestWithContext(r.Context(), "GET", wsURL, nil)
-	if err != nil {
-		log.Printf("[PUBLIC-FILE] Failed to create request: %v", err)
-		http.Error(w, "internal error", http.StatusInternalServerError)
-		return
-	}
-	req.Header.Set("X-User-ID", uid)
-
-	resp, err := workspaceHTTPClient.Do(req)
-	if err != nil {
-		log.Printf("[PUBLIC-FILE] Workspace request failed: %v", err)
-		http.Error(w, "workspace unavailable", http.StatusBadGateway)
-		return
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != http.StatusOK {
-		body, _ := io.ReadAll(resp.Body)
-		log.Printf("[PUBLIC-FILE] Workspace returned %d: %s", resp.StatusCode, string(body))
-		http.Error(w, "file not found", http.StatusNotFound)
-		return
-	}
-
-	// Proxy content-type and body — force inline display (no download)
-	if ct := resp.Header.Get("Content-Type"); ct != "" {
-		w.Header().Set("Content-Type", ct)
-	}
-	w.Header().Set("Content-Disposition", "inline")
-	io.Copy(w, resp.Body)
-}
-
-// handlePublicFolder lists workspace folder contents via a shareable URL.
-// GET /api/public/folder?path=<base64-encoded-folderpath>
-func (api *StreamingAPI) handlePublicFolder(w http.ResponseWriter, r *http.Request) {
-	encoded := r.URL.Query().Get("path")
-	if encoded == "" {
-		http.Error(w, "path query parameter required", http.StatusBadRequest)
-		return
-	}
-
-	decoded, err := base64.StdEncoding.DecodeString(encoded)
-	if err != nil {
-		decoded, err = base64.RawURLEncoding.DecodeString(encoded)
-		if err != nil {
-			log.Printf("[PUBLIC-FOLDER] Failed to decode base64 path: %s, error: %v", encoded, err)
-			http.Error(w, "invalid path encoding", http.StatusBadRequest)
-			return
-		}
-	}
-	folderPath := string(decoded)
-
-	// Use uid from query param (owner's ID for cross-user sharing), fall back to auth context
-	uid := r.URL.Query().Get("uid")
-	if uid == "" {
-		uid = GetUserIDFromContext(r.Context())
-	}
-	log.Printf("[PUBLIC-FOLDER] Listing folder: %s for user: %s", folderPath, uid)
-
-	wsURL := getWorkspaceAPIURL() + "/api/documents"
-	req, err := http.NewRequestWithContext(r.Context(), "GET", wsURL, nil)
-	if err != nil {
-		http.Error(w, "internal error", http.StatusInternalServerError)
-		return
-	}
-	q := req.URL.Query()
-	q.Set("folder", folderPath)
-	req.URL.RawQuery = q.Encode()
-	req.Header.Set("X-User-ID", uid)
-
-	resp, err := workspaceHTTPClient.Do(req)
-	if err != nil {
-		log.Printf("[PUBLIC-FOLDER] Workspace request failed: %v", err)
-		http.Error(w, "workspace unavailable", http.StatusBadGateway)
-		return
-	}
-	defer resp.Body.Close()
-
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(resp.StatusCode)
-	io.Copy(w, resp.Body)
-}
-
-// handlePublicFolderDownload exports a shared folder as a ZIP download.
-// GET /api/public/folder/download?path=<base64-encoded-folderpath>&uid=<owner-id>
-func (api *StreamingAPI) handlePublicFolderDownload(w http.ResponseWriter, r *http.Request) {
-	encoded := r.URL.Query().Get("path")
-	if encoded == "" {
-		http.Error(w, "path query parameter required", http.StatusBadRequest)
-		return
-	}
-
-	decoded, err := base64.StdEncoding.DecodeString(encoded)
-	if err != nil {
-		decoded, err = base64.RawURLEncoding.DecodeString(encoded)
-		if err != nil {
-			http.Error(w, "invalid path encoding", http.StatusBadRequest)
-			return
-		}
-	}
-	folderPath := string(decoded)
-
-	uid := r.URL.Query().Get("uid")
-	if uid == "" {
-		uid = GetUserIDFromContext(r.Context())
-	}
-	log.Printf("[PUBLIC-FOLDER-DOWNLOAD] Exporting folder: %s for user: %s", folderPath, uid)
-
-	// Proxy to workspace export endpoint
-	wsURL := getWorkspaceAPIURL() + "/api/workspace/export"
-	body, _ := json.Marshal(map[string]string{"workspace_path": folderPath})
-	req, err := http.NewRequestWithContext(r.Context(), "POST", wsURL, bytes.NewReader(body))
-	if err != nil {
-		http.Error(w, "internal error", http.StatusInternalServerError)
-		return
-	}
-	req.Header.Set("Content-Type", "application/json")
-	req.Header.Set("X-User-ID", uid)
-
-	client := &http.Client{Timeout: 120 * time.Second}
-	resp, err := client.Do(req)
-	if err != nil {
-		log.Printf("[PUBLIC-FOLDER-DOWNLOAD] Workspace request failed: %v", err)
-		http.Error(w, "workspace unavailable", http.StatusBadGateway)
-		return
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != http.StatusOK {
-		respBody, _ := io.ReadAll(resp.Body)
-		log.Printf("[PUBLIC-FOLDER-DOWNLOAD] Workspace returned %d: %s", resp.StatusCode, string(respBody))
-		http.Error(w, "export failed", resp.StatusCode)
-		return
-	}
-
-	// Proxy ZIP response headers and body
-	for _, h := range []string{"Content-Type", "Content-Disposition", "Content-Length"} {
-		if v := resp.Header.Get(h); v != "" {
-			w.Header().Set(h, v)
-		}
-	}
-	io.Copy(w, resp.Body)
-}
-
-// API Key Validation endpoint - validates API keys for supported providers.
-// Capabilities endpoint
-func (api *StreamingAPI) handleCapabilities(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(http.StatusOK)
-
-	// Get current tracing provider
-	tracingProvider := os.Getenv("TRACING_PROVIDER")
-	if tracingProvider == "" {
-		tracingProvider = "noop"
-	}
-
-	json.NewEncoder(w).Encode(map[string]interface{}{
-		"providers":   []string{"bedrock", "openai", "anthropic"},
-		"streaming":   true,
-		"sse":         true,
-		"agent_modes": []string{"multi-agent", "workflow"},
-		"tracing": map[string]interface{}{
-			"enabled":  tracingProvider != "noop",
-			"provider": tracingProvider,
-		},
-		"workspace":  map[string]interface{}{},
-		"servers":    []string{},
-		"local_mode": IsLocalMode(),
-	})
-}
-
-// handleCdpCheck checks if a CDP port is reachable on localhost
-func (api *StreamingAPI) handleCdpCheck(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-
-	portStr := r.URL.Query().Get("port")
-	if portStr == "" {
-		portStr = "9222"
-	}
-
-	port, err := strconv.Atoi(portStr)
-	if err != nil || port < 1 || port > 65535 {
-		w.WriteHeader(http.StatusBadRequest)
-		json.NewEncoder(w).Encode(map[string]interface{}{
-			"connected": false,
-			"error":     "invalid port number",
-		})
-		return
-	}
-
-	conn, err := net.DialTimeout("tcp", fmt.Sprintf("127.0.0.1:%d", port), 2*time.Second)
-	if err != nil {
-		w.WriteHeader(http.StatusOK)
-		json.NewEncoder(w).Encode(map[string]interface{}{
-			"connected": false,
-		})
-		return
-	}
-	conn.Close()
-
-	w.WriteHeader(http.StatusOK)
-	json.NewEncoder(w).Encode(map[string]interface{}{
-		"connected": true,
-	})
 }
 
 // Query endpoint - handles POST requests to start agent streaming
